% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ac_1}
\alias{ac_1}
\title{Calculate autocorrelation function at lag 1}
\usage{
ac_1(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate autocorrelation function at lag 1
}
\examples{
x <- rnorm(100)
ac_1(x)

}
\author{
Trent Henderson
}
