% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ac_2}
\alias{ac_2}
\title{Calculate autocorrelation function at lag 2}
\usage{
ac_2(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate autocorrelation function at lag 2
}
\examples{
x <- rnorm(100)
ac_2(x)

}
\author{
Trent Henderson
}
