% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_descriptions}
\alias{feature_descriptions}
\title{All features available in tsCeatures in tidy format}
\format{
A dataframe with 2 variables:
\describe{
  \item{names}{Name of the feature}
  \item{description}{Description of what the feature represents}
}
}
\usage{
feature_descriptions
}
\description{
The variables include:
}
\keyword{datasets}
