% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_features}
\alias{get_features}
\title{Automatically run every time-series feature calculation in the package}
\usage{
get_features(x)
}
\arguments{
\item{x}{a numerical time-series vector}
}
\value{
object of class DataFrame that contains the summary statistics for each feature
}
\description{
Automatically run every time-series feature calculation in the package
}
\examples{
x <- rnorm(100)
outs <- get_features(x)

}
\author{
Trent Henderson
}
