% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate tailedness of the value distribution}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate tailedness of the value distribution
}
\examples{
x <- rnorm(100)
kurtosis(x)

}
\author{
Trent Henderson
}
