% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prop_above_mean}
\alias{prop_above_mean}
\title{Calculate proportion of points above the mean, as per binarize_mean from the MATLAB toolbox \code{hctsa}}
\usage{
prop_above_mean(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate proportion of points above the mean, as per binarize_mean from the MATLAB toolbox \code{hctsa}
}
\examples{
x <- rnorm(100)
prop_above_mean(x)

}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).
}
\author{
Trent Henderson
}
