% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_10}
\alias{quantile_10}
\title{Calculate Quantile 10}
\usage{
quantile_10(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate Quantile 10
}
\examples{
x <- rnorm(100)
quantile_10(x)

}
\author{
Trent Henderson
}
