% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_25}
\alias{quantile_25}
\title{Calculate Quantile 25}
\usage{
quantile_25(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate Quantile 25
}
\examples{
x <- rnorm(100)
quantile_25(x)

}
\author{
Trent Henderson
}
