% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_5}
\alias{quantile_5}
\title{Calculate Quantile 5}
\usage{
quantile_5(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate Quantile 5
}
\examples{
x <- rnorm(100)
quantile_5(x)

}
\author{
Trent Henderson
}
