% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_75}
\alias{quantile_75}
\title{Calculate Quantile 75}
\usage{
quantile_75(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate Quantile 75
}
\examples{
x <- rnorm(100)
quantile_75(x)

}
\author{
Trent Henderson
}
