% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_90}
\alias{quantile_90}
\title{Calculate Quantile 90}
\usage{
quantile_90(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate Quantile 90
}
\examples{
x <- rnorm(100)
quantile_90(x)

}
\author{
Trent Henderson
}
