% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_95}
\alias{quantile_95}
\title{Calculate Quantile 95}
\usage{
quantile_95(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate Quantile 95
}
\examples{
x <- rnorm(100)
quantile_95(x)

}
\author{
Trent Henderson
}
