% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skewness}
\alias{skewness}
\title{Calculate adjusted Fisher-Pearson coefficient of skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate adjusted Fisher-Pearson coefficient of skewness
}
\examples{
x <- rnorm(100)
skewness(x)

}
\author{
Trent Henderson
}
