% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{walker_propcross}
\alias{walker_propcross}
\title{Calculate proportion of length that walker crosses time series, as per PH_Walker_prop_01_sw_propcross from the MATLAB toolbox \code{hctsa}}
\usage{
walker_propcross(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate proportion of length that walker crosses time series, as per PH_Walker_prop_01_sw_propcross from the MATLAB toolbox \code{hctsa}
}
\examples{
x <- rnorm(100)
walker_propcross(x)

}
\references{
B.D. Fulcher and N.S. Jones. hctsa: A computational framework for automated time-series phenotyping using massive feature extraction. Cell Systems 5, 527 (2017).

B.D. Fulcher, M.A. Little, N.S. Jones Highly comparative time-series analysis: the empirical structure of time series and their methods. J. Roy. Soc. Interface 10, 83 (2013).

Rob Hyndman, Yanfei Kang, Pablo Montero-Manso, Thiyanga Talagala, Earo Wang, Yangzhuoran Yang and Mitchell O'Hara-Wild (2020). tsfeatures: Time Series Feature Extraction. R package version 1.0.2. https://CRAN.R-project.org/package=tsfeatures
}
\author{
Trent Henderson
}
