% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossing_points}
\alias{crossing_points}
\title{Calculate number of crossing points}
\usage{
crossing_points(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value
}
\description{
Calculate number of crossing points
}
\examples{
x <- rnorm(100)
crossing_points(x)

}
\author{
Trent Henderson
}
